Time: 0 ms (100.00%) | Memory: 7.8 MB (68.52%) - LeetSync

+++ 278-first-bad-version/first-bad-version.cpp
@@ -6,7 +6,7 @@ public:
     int firstBadVersion(int n) {
         int ans=-1;
         long low=1,high=n;
-        long long mid;
+        long mid;
 
         while(low<=high)
         {
